SingUp moved out to separate component
